requirements.txt
numpy==1.18.1
six==1.14.0
opencv-python==4.2.0.32
matplotlib
Pillow
